+++ vertx-test/src/main/java/com/cloudentity/tools/vertx/test/ServiceVerticleUnitTest.java
@@ -5,6 +5,9 @@ import com.cloudentity.tools.vertx.bus.ServiceVerticle;
 import io.vertx.core.Future;
 import io.vertx.core.json.JsonObject;
 
+import java.lang.reflect.ParameterizedType;
+import java.lang.reflect.Type;
+
 /**
  * Use it when you want to test ServiceVerticle that doesn't depend on other ServiceVerticles.
  * Otherwise use {@link ServiceVerticleIntegrationTest}
@@ -19,7 +22,23 @@ abstract public class ServiceVerticleUnitTest<B extends ServiceVerticle, A> exte
   }
 
   public A client() {
-    return (A) VertxEndpointClient.make(vertx(), createVerticle().vertxServices().get(0));
+    return (A) VertxEndpointClient.make(vertx(), getGenericClass());
+  }
+
+  private Class<A> inferedClass;
+
+  public Class<A> getGenericClass(){
+    if(inferedClass == null){
+      Type mySuperclass = getClass().getGenericSuperclass();
+      Type tType = ((ParameterizedType)mySuperclass).getActualTypeArguments()[1];
+      String className = tType.toString().split(" ")[1];
+      try {
+        inferedClass = (Class<A>) Class.forName(className);
+      } catch (ClassNotFoundException e) {
+        e.printStackTrace();
+      }
+    }
+    return inferedClass;
   }
 
   public Future<String> deployVerticle() {